requirements.txt
aerich==0.3.3
aiohttp==3.6.3
aiosqlite==0.16.0
async-timeout==3.0.1
asyncpg==0.21.0
atlastk==0.12.1
attrs==19.3.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
discord.py==1.5.1
idna==2.10
iso8601==0.1.13
loguru==0.5.3
multidict==4.7.6
packaging==20.4
pydantic==1.7.2
PyPika==0.44.0
tortoise-orm==0.16.17
typing-extensions==3.7.4.3
win32-setctime==1.0.3
yarl==1.5.1
